% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{adequacy}
\alias{adequacy}
\title{adequacy
 Calculate the summary of the five method above.}
\usage{
adequacy(sol, Y, A, X = NULL, nb = 10000, alpha = 0.98)
}
\arguments{
\item{sol}{a object of type trajectory.}

\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{an optionnal matrix that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}
}
\value{
a table of real. A table with 5 rows: the estimate probabilities, the two bounds
of the confidence interval, the proportion of assignement, the Average Posterior Probability and
the Odds of Correct Classifification.
}
\description{
adequacy
 Calculate the summary of the five method above.
}
