% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{AvePP}
\alias{AvePP}
\title{AvePP
Calculate the Average Posterior Porbability.
 Average Posterior Probability (AvePP) the average posterior
 probability of membership for each group for those individuals that were assigned to.}
\usage{
AvePP(sol, Y, A, X = NULL)
}
\arguments{
\item{sol}{a object of type trajectory.}

\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{an optionnal matrix that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}
}
\value{
a vector of real. The average posterior probability.
}
\description{
AvePP
Calculate the Average Posterior Porbability.
 Average Posterior Probability (AvePP) the average posterior
 probability of membership for each group for those individuals that were assigned to.
}
