% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{ConfIntT}
\alias{ConfIntT}
\title{ConfIntT
 Calculate the confidence interval of the probibiliies with a bootstrap method. We have to speicify the number of the repetition
 of bootstrap and the degree of confidence.}
\usage{
ConfIntT(sol, Y, A, X = NULL, nb = 10000, alpha = 0.98)
}
\arguments{
\item{sol}{a object of type trajectory.}

\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{an optionnal matrix that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}
}
\value{
a vector of real. The two bounds of the confidence interval given a degree of confidence.
}
\description{
ConfIntT
 Calculate the confidence interval of the probibiliies with a bootstrap method. We have to speicify the number of the repetition
 of bootstrap and the degree of confidence.
}
