% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataLOGIT01}
\alias{dataLOGIT01}
\title{Data for Logit's example}
\format{
A data frame with 500 rows :
\describe{
  \item{Y}{time values of individual}
  \item{T}{time values}
  \item{X}{porbability's prescriptors}
  \item{V}{time dependent covariate}
}
}
\usage{
data(dataLOGIT01)
}
\description{
An artificial dataset.
}
\keyword{datasets}
