% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataNORM01}
\alias{dataNORM01}
\title{Data for censored normal's example}
\format{
A data frame with 300 rows :
\describe{
  \item{Y}{time values of individual}
  \item{T}{time values}
  \item{X}{porbability's prescriptors}
  \item{gr}{group's membership}
}
}
\usage{
data(dataNORM01)
}
\description{
An artificial dataset.
}
\keyword{datasets}
