% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NonLinear.R
\name{fait}
\alias{fait}
\title{Function fait}
\usage{
fait(betak, i, t, A, TCOV, fct, diffct)
}
\arguments{
\item{betak}{Vector of integer.}

\item{i}{Integer.}

\item{t}{Real.}

\item{A}{Matrix of real.}

\item{TCOV}{Matrix of real.}

\item{fct}{Function.}

\item{diffct}{Function.}
}
\value{
real. Compute the value of the function fct for individual i, time t and group k.
}
\description{
Function fait
}
