% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.R
\name{ftauxLogit}
\alias{ftauxLogit}
\title{Calculate the tau in EM algorithm for a Logit model.}
\usage{
ftauxLogit(pi, beta, ng, n, nbeta, A, Y, TCOV, delta, nw, nx, X)
}
\arguments{
\item{pi}{A vector of probability.}

\item{beta}{A vector.}

\item{ng}{a interger.}

\item{n}{a integer.}

\item{nbeta}{a vector of integer.}

\item{A}{a matrix of real which is time index.}

\item{Y}{a matrix of real which is ?.}
}
\value{
A vector wich correspond to tau.
}
\description{
Calculate the tau in EM algorithm for a Logit model.
}
