% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Censorednormal.R
\name{gkCNORM}
\alias{gkCNORM}
\title{Title}
\usage{
gkCNORM(beta, sigma, i, k, nbeta, A, Y, ymin, ymax, TCOV, delta, nw)
}
\arguments{
\item{beta}{vector of real. The values of beta parameters.}

\item{sigma}{a vector of reals. The values of sigma.}

\item{nbeta}{vector of integer. The number of parameters beta for each group.}

\item{A}{a matrix containing the time variable data.}

\item{Y}{a matrix containing the variables in the model.}

\item{ymin}{a real. For censored data, it is the minimum value of the data.}

\item{ymax}{a real. For censored data, it is the maxium value of the data.}

\item{TCOV}{a matrix containing the time covariate that influence the trajectory themselve.
By default its value is NULL.}

\item{delta}{vector of real. The values of delta parameters.}

\item{nw}{integer. The number of time dependant variable. 0 if TCOV is NULL.}
}
\value{
a real. The porduct of probability for each group.
}
\description{
Title
}
