% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunction.R
\name{GroupProb}
\alias{GroupProb}
\title{GroupProb calculate the membership probability of each value of the data.}
\usage{
GroupProb(Obj, Y, A, TCOV = NULL, X = NULL)
}
\arguments{
\item{Obj}{a trajectory object that is return by trajeR function.}

\item{Y}{a real matrix. The data.}

\item{A}{a real matrix. The time variable.}

\item{TCOV}{a real matrix. Optionnal, by default the value is NULL. It contained the time dependent covariate.}

\item{X}{a real matrix. Optionnal, by default the value is NULL. It contained a caovraite that modify the probaility memebrship.}
}
\value{
a real matrix. For each individual i in the data, this matrix contained the membership probability of each group.
}
\description{
GroupProb calculate the membership probability of each value of the data.
}
\examples{
GroupProb(solL, Y=data[,2:11], A=data[,11:21])
}
