% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommonFunction.R
\name{Likelihood}
\alias{Likelihood}
\title{Likelihood of an trajectory object}
\usage{
Likelihood(
  param,
  model,
  method,
  ng,
  nx,
  n,
  nbeta,
  nw,
  A,
  Y,
  X,
  TCOV,
  ymin = NULL,
  ymax = NULL,
  nnu = NULL,
  fct = NULL
)
}
\arguments{
\item{param}{a vector with the parameters of the model.}

\item{model}{a string with the name of the model.}

\item{ng}{a integer. The number of groups.}

\item{nx}{a integer. The number of covariate that influenced the belonging probability.}

\item{n}{a integer. The number of variable in the data.}

\item{nbeta}{a vector of intger. This vector contained the degre of each polynomial shape of trajectory.}

\item{nw}{a integer. The number of covariate that influenced the shape of trajectory.}

\item{A}{a matrix of real. The time variable.}

\item{Y}{a matrix of real. The data.}

\item{X}{a matrix of real. The values that influenced the belonging probability.
By default thsi matrxi is a matrix with 1.}

\item{TCOV}{a matrix of real. The values that influenced the shape of trajectory.
By default this matrrix is NULL.}

\item{ymin}{real. For the CNORM model, indicate the minimum value of the data. It oncerne only the model
with censored data. By default its value is NULL.}

\item{ymax}{real. For the CNORM model, indicate the maximum value of the data. It oncerne only the model
with censored data. By default its value is NULL.}

\item{nnu}{a vector of real. The degre of the Poisson part.}
}
\value{
a real. The Likelihood of the mixture model.
}
\description{
Calculate the likelihood of an object with a particular trajectory class.
}
\examples{
Likelihood(param, model, ng, nx, n, nbeta, nw, A, Y, X, TCOV)
}
