% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{piik_cpp}
\alias{piik_cpp}
\title{piik_cpp}
\usage{
piik_cpp(theta, i, k, ng, X)
}
\arguments{
\item{theta}{a vector of real with the value of theta parameters.}

\item{i}{a integer. The index of the data which is used to calculate the probability of group belonging.}

\item{k}{a integer. The numerous of group.}

\item{ng}{a integer. The number of groups.}

\item{X}{a matrix. The covariate that influence the group belonging probability.}
}
\value{
A vector of real. The probability for an individual to belonging at each group.
}
\description{
Calculate the membership probability to group k for a individual i.
}
