% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrajeR.R
\name{plot.Trajectory.CNORM}
\alias{plot.Trajectory.CNORM}
\title{plot CNORM trajectory}
\usage{
\method{plot}{Trajectory.CNORM}(
  Obj,
  plotcov = NULL,
  col = "black",
  Y = NULL,
  A = NULL,
  X = NULL,
  mean = FALSE,
  alpha = 1,
  ...
)
}
\arguments{
\item{plotcov}{an optionnal vector or matrix with the same length as the time period. Default value is NULL.}

\item{col}{an optionnal vector. The vecotr of colors. It must contain a color for each trajectory and each points of groups.
Its length is the double of the number of group. Default valme is a grayscale.}

\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{an optionnal matrix that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}

\item{mean}{an optional logicial. Indicate if the mean of ech group and time value must be draw.}

\item{alpha}{on optionnal real. Indiciate the alpha channel of the points color.}

\item{...}{}

\item{dec}{an optionnal real. It precise the shift to draw the data points.}
}
\value{
a graphic.
}
\description{
#' @param Obj an object of class "\code{Trajectory.LOGIT}".
}
\examples{
plot(solL)
}
