% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrajeR.R
\name{plot.Trajectory.ZIP}
\alias{plot.Trajectory.ZIP}
\title{plot ZIP trajectory}
\usage{
\method{plot}{Trajectory.ZIP}(
  Obj,
  plotcov = NULL,
  dec = 1,
  col = "black",
  Y = NULL,
  A = NULL,
  X = NULL,
  TCOV = NULL,
  mean = FALSE,
  alpha = 1,
  ...
)
}
\description{
plot ZIP trajectory
}
