% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrajeR.R
\name{plotrajeR}
\alias{plotrajeR}
\title{plot trajectory}
\usage{
plotrajeR(Obj, ...)
}
\arguments{
\item{Obj}{an object of class "\code{Trajectory.CNORM}".}

\item{...}{optional parameters}
}
\value{
a graphic.
}
\description{
plot trajectory
}
\examples{
\dontrun{
plotrajeR(solL)
}
}
