% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.BETA}
\alias{print.Trajectory.BETA}
\title{Print BETA}
\usage{
\method{print}{Trajectory.BETA}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. An object of class "\code{Trajectory.BETA}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print method for an object of class "\code{Trajectory.BETA}".
}
