% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.CNORM}
\alias{print.Trajectory.CNORM}
\title{Print CNORM}
\usage{
\method{print}{Trajectory.CNORM}(Obj, ...)
}
\arguments{
\item{Obj}{an object of class "\code{Trajectory.CNORM}".}

\item{...}{}
}
\value{
The print of Obj.
}
\description{
Print mehtod for an object of class "\code{Trajectory.CNORM}".
}
\examples{
load(file='data/CNORM_data01')
solL
}
