% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.LOGIT}
\alias{print.Trajectory.LOGIT}
\title{Print LOGIT}
\usage{
\method{print}{Trajectory.LOGIT}(Obj, ...)
}
\arguments{
\item{Obj}{an object of class "\code{Trajectory.LOGIT}".}

\item{...}{}
}
\value{
The print of Obj.
}
\description{
Print mehtod for an object of class "\code{Trajectory.LOGIT}".
}
\examples{
load(file='data/LOGIT_data01')
solL
}
