% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.NL}
\alias{print.Trajectory.NL}
\title{print NL trajectory}
\usage{
\method{print}{Trajectory.NL}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. . An object of class "\code{Trajectory.NL}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print method for an object of class "\code{Trajectory.NL}".
}
