% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.POIS}
\alias{print.Trajectory.POIS}
\title{Print POIS}
\usage{
\method{print}{Trajectory.POIS}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. . An object of class "\code{Trajectory.POIS}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print mehtod for an object of class "\code{Trajectory.POIS}".
}
