% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.ZIP}
\alias{print.Trajectory.ZIP}
\title{Print ZIP}
\usage{
\method{print}{Trajectory.ZIP}(Obj, ...)
}
\value{
The print of Obj.
}
\description{
Print method for an object of class "\code{Trajectory.ZIP}".
}
