% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{propAssign}
\alias{propAssign}
\title{propAssign
 Calculate the proportion of individuals in a given group. That is the ratio of the number of
 individuals in one group and all the individuals.}
\usage{
propAssign(sol, Y, A, X = NULL)
}
\arguments{
\item{sol}{a object of type trajectory.}

\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{an optionnal matrix that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}
}
\value{
a vector of real. The  proportion.
}
\description{
propAssign
 Calculate the proportion of individuals in a given group. That is the ratio of the number of
 individuals in one group and all the individuals.
}
