% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalTrajeR.R
\name{trajeR.CNORM}
\alias{trajeR.CNORM}
\title{Title}
\usage{
trajeR.CNORM(
  Y,
  A,
  X,
  TCOV,
  ng,
  nx,
  n,
  nbeta,
  nw,
  ntheta,
  period,
  degre,
  theta,
  beta,
  sigma,
  delta,
  pi,
  Method,
  ssigma,
  ymax,
  ymin,
  hessian,
  itermax,
  paraminit,
  EMIRLS,
  refgr
)
}
\arguments{
\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{data that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}

\item{TCOV}{a matrix containing the time covariate that influence the trajectory themselve.
By default its value is NULL.}

\item{ng}{integer. The number of group.}

\item{nx}{integer. The number ov variable X.}

\item{n}{integer. The number of individuals. It is the number of row of Y too.}

\item{nbeta}{vector of integer. The number of parameters beta for each group.}

\item{nw}{integer. The number of time dependant variable. 0 if TCOV is NULL.}

\item{ntheta}{vector of integer. The number of covariate X plus 1.}

\item{period}{integer. the number of time period.}

\item{degre}{vector of integer. The degre of the ploynomial shape.}

\item{theta}{vector of real. The values of theta parameters.}

\item{beta}{vector of real. The values of beta parameters.}

\item{sigma}{a vector of reals. The values of sigma.}

\item{delta}{vector of real. The values of delta parameters.}

\item{pi}{vector of real. The values of pi parameters.}

\item{Method}{string. The method used, Likelihhod, Expectation Maximization or
Expectation Maximization IRWLS.}

\item{ssigma}{logical. A boolean thaht indicates if we had to use same sigma for all groups. By default is FALSE.}

\item{ymax}{a real. For censored data, it is the maxium value of the data.}

\item{ymin}{a real. For censored data, it is the minimum value of the data.}

\item{hessian}{logicial. If TRUE the hessian matrix is calculated. By default its value is FALSE.}

\item{itermax}{integer. Indicate the maximal number of iteration for optim function or for the EM algorithm.}

\item{paraminit}{vector. The vector of initial parameters. By default trajeR calculate the initial value
based of the range or the standrad deviation.}

\item{EMIRLS}{logicial. If TRUE  and if we use predictors for the membership probability we fit the partametrs
by using IRLS method int he EM algorithm. By default its value is FALSE.}

\item{refgr}{integer. Indicate the group reference for the calculus of the membership probability.}
}
\description{
Title
}
