% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InternalTrajeR.R
\name{trajeR.LOGIT}
\alias{trajeR.LOGIT}
\title{trajeR.LOGIT}
\usage{
trajeR.LOGIT(
  Y,
  A,
  X,
  TCOV,
  ng,
  nx,
  n,
  nbeta,
  nw,
  ntheta,
  period,
  degre,
  theta,
  beta,
  delta,
  pi,
  Method,
  hessian,
  itermax,
  paraminit,
  EMIRLS,
  refgr
)
}
\arguments{
\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{data that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}

\item{TCOV}{a matrix containing the time covariate that influence the trajectory themselve.
By default its value is NULL.}

\item{ng}{integer. The number of group.}

\item{nx}{integer. The number ov variable X.}

\item{n}{integer. The number of individuals. It is the number of row of Y too.}

\item{nbeta}{vector of integer. The number of parameters beta for each group.}

\item{nw}{integer. The number of time dependant variable. 0 if TCOV is NULL.}

\item{ntheta}{vector of integer. The number of covariate X plus 1.}

\item{period}{integer. the number of time period.}

\item{degre}{vector of integer. The degre of the ploynomial shape.}

\item{theta}{vector of real. The values of theta parameters.}

\item{beta}{vector of real. The values of beta parameters.}

\item{delta}{vector of real. The values of delta parameters.}

\item{pi}{vector of real. The values of pi parameters.}

\item{Method}{string. The method used, Likelihhod, Expectation Maximization or
Expectation Maximization IRWLS.}

\item{hessian}{logicial. If TRUE the hessian matrix is calculated. By default its value is FALSE.}

\item{itermax}{integer. Indicate the maximal number of iteration for optim function or for the EM algorithm.}

\item{paraminit}{vector. The vector of initial parameters. By default trajeR calculate the initial value
based of the range or the standrad deviation.}

\item{EMIRLS}{logicial. If TRUE  and if we use predictors for the membership probability we fit the partametrs
by using IRLS method int he EM algorithm. By default its value is FALSE.}

\item{refgr}{integer. Indicate the group reference for the calculus of the membership probability.}
}
\value{
return a list object.
\itemize{
  \item beta -  vector of the paramters beta.
  \item delta - vector of the parameter delta. Only if we use time covariate.
  \item theta - vector with the paramter theta if there exist a coavriate X that modifie
  the probability or the probability of group membership.
  \item sd - vector of the standrad deviation of the parameters.
  \item tab - a matrix with all the parameters and standard deviation.
  \item Likelihoo -  a real with the Likelihhod obtnaied by the parameters.
  \item ng - a integer with the number of group.
  \item model - a string with the model used.
  \item method -  a string with the method used.
}
}
\description{
trajeR.LOGIT
}
