% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{trajeR}
\alias{trajeR}
\title{Fitting longitudinal mixture models}
\usage{
trajeR(
  Y,
  A,
  Risk = NULL,
  TCOV = NULL,
  ng,
  degre = NULL,
  degre.nu = 0,
  Model,
  Method = "L",
  ssigma = FALSE,
  ymax = max(Y) + 1,
  ymin = min(Y) - 1,
  hessian = TRUE,
  itermax = 100,
  paraminit = NULL,
  EMIRLS = TRUE,
  refgr = 1,
  fct = NULL,
  diffct = NULL,
  nbvar = NULL,
  nls.lmiter = 50
)
}
\arguments{
\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{TCOV}{an optionnal matrix containing the time covariate that influence the trajectory themselve.
By default its value is NULL.}

\item{ng}{integer. The number of group.}

\item{degre}{vector of integer. The degre of every ploynomial function.}

\item{degre.nu}{vecor of integer. The degre of all Poisson part for a ZIP model.}

\item{Model}{the model used. The value are LOGIT for a Logit Mixture model,
CNORM for a Censored Normal Mixture Model or ZIP for Zero Inflated Poisson Mixture model.}

\item{Method}{determine the method used for find the paramters of the model.
The value are L for the Maximum Likelihiood Estimation, EM for Expectation Maximization method
with quasi newton method inside, EMIWRLS for Expectation Maximization method with Iterative
Weighted ??}

\item{ssigma}{logical. By default its value is FALSE. For the CNORM model,
indicate if we want the same sigma for all nomrmal density function.}

\item{ymax}{real. For the CNORM model, indicate the maximum value of the data. It oncerne only the model
with censored data. By default its value is the maximum value of the data plus 1.}

\item{ymin}{real. For the CNORM model, indicate the minimum value of the data. It oncerne only the model
with censored data. By default its value is the maximum value of the data minus 1.}

\item{hessian}{logical. Indicate if we want calculate the hessian matrix. Default is FALSE.
If the method use is Likelihood, the hessian is calculated by inversing the Information's Fisher Matrix.
To avoid numerically singular matrix we find the pseudo inverse matrix by using the ginv function int he package MASS.
If the method is EM or EMIWRLS, the hessian is calculted by using Louis method.}

\item{itermax}{integer. Indicate the maximal number of iteration for optim function or for the EM algorithm.}

\item{paraminit}{vector. The vector of initial parameters. By default trajeR calculate the initial value
based of the range or the standrad deviation.}

\item{fct}{a function. The definition of the function  f in the definition in nonlinear model.}

\item{diffct}{a function. The differential of the function f in the nonlinear model.}

\item{nbvar}{integer. The number of variable in the nonlinear model.}

\item{X}{an optionnal matrix that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}
}
\value{
return an object of class "\code{Trajectory.LOGIT}".
The generic accessor functions \code{beta}, \code{delta}, \code{theta}, \code{sd}, \code{tab}, \code{Likelihood},
\code{ng}, \code{model} and \code{method} extract various useful features of the value returned by \code{trajeR}.

An object of class "\code{Trajectory.LOGIT}" is a list containing at least the following components:

\describe{
\item{\code{beta}}{a vector of the parameters beta.}
\item{\code{delta}}{a vector of the parameter delta. Only if we use time covariate.}
\item{\code{theta}}{a vector with the paramter theta if there exist a coavriate X that modifie
  the probability or the probability of group membership.}
\item{\code{sd}}{a vector of the standrad deviation of the parameters.}
\item{\code{tab}}{a matrix with all the parameters and standard deviation.}
\item{\code{Likelihood}}{a real with the Likelihhod obtnaied by the parameters.}
\item{\code{ng}}{a integer with the number of group.}
\item{\code{model}}{a string with the model used.}
\item{\code{method}}{a string with the method used.}
}
}
\description{
\code{trajeR} is used to fit longitudinal mixture models. It used 3 types of mixture models : LOGIT, ZIP and censored Normal.
}
\details{
Models for trajeR is, by default, a polynomial regression of the time value parameters for each groups. The number fo group is controlled by the integer \code{ng}.
We can spcecify the degre of the polynomial shape for each groups by the vector \code{degre}.
}
\examples{
load(file = "/data/LOGIT_data01")
data=as.matrix(data)
solL = trajeR(data[,2:11], data[,12:21], ng = 3, degre=c(0,3,4),
              Model="LOGIT", Method = "L", hessian = TRUE,
              itermax = 300, paraminit = paraminit)
}
