% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajeRmodelSelection.R
\name{trajeRAIC}
\alias{trajeRAIC}
\title{AIC function to an trajectory object}
\usage{
trajeRAIC(sol)
}
\arguments{
\item{sol}{Trajectory's object. An object of type trajectory.}
}
\value{
A real.
}
\description{
Calculate the AIC value to an trajectory object.
}
\examples{
\dontrun{
load("data/dataNORM01.RData")
solL = trajeR(data[,1:5], data[,6:10], ng = 3, degre=c(2,2,2), 
              Model="CNORM", Method = "L", ssigma = FALSE, 
              hessian = TRUE)
trajeAIC(solL)
}
}
