% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajeRmodelSelection.R
\name{trajeRSH}
\alias{trajeRSH}
\title{SH function to an trajectory object}
\usage{
trajeRSH(l)
}
\arguments{
\item{l}{List. A list of objects of type trajectory.}
}
\value{
A vector of real.
}
\description{
Calculate the Slope Heuristic value to a list of trajectory objects.
}
\examples{
\dontrun{
load("data/dataNORM01.RData")
solL = trajeR(data[,1:5], data[,6:10], ng = 3, degre=c(2,2,2), 
              Model="CNORM", Method = "L", ssigma = FALSE, 
              hessian = TRUE)
trajeSH(solL)
}
}
