% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{adequacy}
\alias{adequacy}
\title{Adequacy}
\usage{
adequacy(sol, Y, A, X = NULL, nb = 10000, alpha = 0.98)
}
\arguments{
\item{sol}{An object of type trajectory.}

\item{Y}{A matrix containing the variables in the model.}

\item{A}{A matrix containing the time variable data.}

\item{X}{An optionnal matrix that modifie the probabilty of belong to group.
By default its value is a mtrix with one column  with value 1.}

\item{nb}{An integer. The numbers of repetition in the bootstrap method.}

\item{alpha}{A number. The degree of confidence of the interval.}
}
\value{
A table of real. A table with 5 rows: the estimate probabilities, the
 two bounds of the confidence interval, the proportion of assignment, the
 Average Posterior Probability and the Odds of Correct Classification.
}
\description{
Calculate the summary of the five method above.
}
\examples{
load("data/dataNORM01.RData")
sol = trajeR(data[,1:5], data[,6:10], ng = 3, degre=c(2,2,2), Model="CNORM",
Method = "L", ssigma = FALSE, hessian = FALSE)
adequacy (sol, Y = data[, 1;5], A = data[, 6:10])
}
