% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{AvePP}
\alias{AvePP}
\title{Average Posterior Probability}
\usage{
AvePP(sol, Y, A, X = NULL)
}
\arguments{
\item{sol}{Trajectory's object. An object of type trajectory.}

\item{Y}{Matrix. A matrix containing the variables in the model.}

\item{A}{Matrix. A matrix containing the time variable data.}

\item{X}{Matrix. An optional matrix that modify the probability of belong to group.
By default its value is a matrix with one column  with value 1.}
}
\value{
A vector of real. The average posterior probability.
}
\description{
Calculate the Average Posterior Probability. Average Posterior Probability
(AvePP) the average posterior probability of membership for each group for
those individuals that were assigned to.
}
\examples{
\dontrun{
load("data/dataNORM01.RData")
solL = trajeR(data[,1:5], data[,6:10], ng = 3, degre=c(2,2,2), 
              Model="CNORM", Method = "L", ssigma = FALSE, 
              hessian = TRUE)
AvePP(sol, Y = data[, 1:5], A = data[, 6:10])
}
}
