% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{ConfIntT}
\alias{ConfIntT}
\title{Confidence interval}
\usage{
ConfIntT(sol, Y, A, X = NULL, nb = 10000, alpha = 0.98)
}
\arguments{
\item{sol}{An object of type trajectory.}

\item{Y}{A matrix containing the variables in the model.}

\item{A}{A matrix containing the time variable data.}

\item{X}{An optionnal matrix that modifie the probabilty of belong to group.
By default its value is a matrix with one column  with value 1.}

\item{nb}{An integer. The numbers of repetition in the bootstrap method.}

\item{alpha}{A number. The degree of confidence of the interval.}
}
\value{
A vector of real. The two bounds of the confidence interval given a
  degree of confidence.
}
\description{
Calculate the confidence interval of the probabilities with a bootstrap
method. We have to specify the number of the repetition of bootstrap and the
degree of confidence.
}
\examples{
load("data/dataNORM01.RData")
solL = trajeR(data[,1:5], data[,6:10], ng = 3, degre=c(2,2,2), 
              Model="CNORM", Method = "L", ssigma = FALSE, 
              hessian = TRUE)
ConfIntT(sol, Y = data[, 1;5], A = data[, 6:10])
}
