% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupProfiles.R
\name{GroupProfiles}
\alias{GroupProfiles}
\title{GroupProfiles}
\usage{
GroupProfiles(sol, Y, A, X)
}
\arguments{
\item{sol}{a object of type trajectory.}

\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{an optionnal matrix that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}
}
\value{
a table of real.
}
\description{
Calculate the group profil regarding covariate. It is a cross tabulation
 of individual level trajectory group assignments with individual
 level characteristic that might be associated with trajectory group membership.
}
