% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GroupProfiles.R
\name{GroupProfiles}
\alias{GroupProfiles}
\title{Profiles of each group}
\usage{
GroupProfiles(sol, Y, A, X)
}
\arguments{
\item{sol}{Trajectory's object. A object of type trajectory.}

\item{Y}{Matirx. A matrix containing the variables in the model.}

\item{A}{Matrix. A matrix containing the time variable data.}

\item{X}{Matrix. An optional matrix that modify the probability of belong to group.
By default its value is a matrix with one column  with value 1.}
}
\value{
A table of real.
}
\description{
\code{GroupProfiles} calculate the profile of a group regarding covariate. It is a
cross tabulation of individual level trajectory group assignments with
individual level characteristic that might be associated with trajectory
group membership.
}
\examples{
\dontrun{
load("data/dataNORM01.RData")
solL = trajeR(data[,1:5], data[,6:10], ng = 3, degre=c(2,2,2), 
              Model="CNORM", Method = "L", ssigma = FALSE, 
              hessian = TRUE)
GroupProfles(sol, Y = data[,1:5], A = data[, 6:10])
}
}
