% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{OCC}
\alias{OCC}
\title{OCC
 Calculate Odds of Correct Classification. The Odds of Correct Classification for
 group k (OCCj) is the ratio between the odds of a correct
 classification into group j on the basis of the posterior probability rule and
 the odds of correct assignment based on random assignments with the probability
 of assignment to group j is done with pik, the probability estimate by
 the model.}
\usage{
OCC(sol, Y, A, X = NULL)
}
\arguments{
\item{sol}{a object of type trajectory.}

\item{Y}{a matrix containing the variables in the model.}

\item{A}{a matrix containing the time variable data.}

\item{X}{an optionnal matrix that modifie the probabilty of belong to group. By default its value is a mtrix
with one column  with value 1.}
}
\value{
a vector of real. The Odds of Correct Classification.
}
\description{
OCC
 Calculate Odds of Correct Classification. The Odds of Correct Classification for
 group k (OCCj) is the ratio between the odds of a correct
 classification into group j on the basis of the posterior probability rule and
 the odds of correct assignment based on random assignments with the probability
 of assignment to group j is done with pik, the probability estimate by
 the model.
}
