% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrajeR.R
\name{plot.Trajectory.NL}
\alias{plot.Trajectory.NL}
\title{plot Non Linear trajectory}
\usage{
\method{plot}{Trajectory.NL}(
  Obj,
  plotcov = NULL,
  col = "black",
  Y = NULL,
  A = NULL,
  X = NULL,
  mean = FALSE,
  alpha = 1,
  main = "Values and predicted trajectories for all groups",
  ...
)
}
\description{
plot Non Linear trajectory
}
