% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotTrajeR.R
\name{plotrajeR}
\alias{plotrajeR}
\title{plot trajectory}
\usage{
plotrajeR(Obj, ...)
}
\arguments{
\item{Obj}{an object of class "\code{Trajectory}".}

\item{...}{optional parameters}
}
\value{
a graphic.
}
\description{
plot trajectory
}
\examples{
data <- read.csv(system.file("extdata", "CNORM2gr.csv", package = "trajeR"))
data <- as.matrix(data)
sol <- trajeR(Y = data[, 2:6], A = data[, 7:11], degre = c(2, 2), Model = "CNORM", Method = "EM")
plotrajeR(sol)

}
