% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.CNORM}
\alias{print.Trajectory.CNORM}
\title{Print CNORM}
\usage{
\method{print}{Trajectory.CNORM}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. An object of class "\code{Trajectory.CNORM}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print method for an object of class "\code{Trajectory.CNORM}".
}
\examples{
\dontrun{
load(file='data/CNORM_data01')
solL
}
}
