% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.CNORM}
\alias{print.Trajectory.CNORM}
\title{Print CNORM}
\usage{
\method{print}{Trajectory.CNORM}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. An object of class "\code{Trajectory.CNORM}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print method for an object of class "\code{Trajectory.CNORM}".
}
\examples{
data <- read.csv(system.file("extdata", "CNORM2gr.csv", package = "trajeR"))
data <- as.matrix(data)
sol <- trajeR(Y = data[, 2:6], A = data[, 7:11], degre = c(2, 2), Model = "CNORM", Method = "EM")
sol
}
