% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.LOGIT}
\alias{print.Trajectory.LOGIT}
\title{Print LOGIT}
\usage{
\method{print}{Trajectory.LOGIT}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. . An object of class "\code{Trajectory.LOGIT}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print mehtod for an object of class "\code{Trajectory.LOGIT}".
}
\examples{
\dontrun{
load(file='data/LOGIT_data01')
solL
}
}
