% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.LOGIT}
\alias{print.Trajectory.LOGIT}
\title{Print LOGIT}
\usage{
\method{print}{Trajectory.LOGIT}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. . An object of class "\code{Trajectory.LOGIT}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print mehtod for an object of class "\code{Trajectory.LOGIT}".
}
\examples{
data = read.csv(system.file("extdata", "LOGIT2gr.csv", package = "trajeR"))
data = as.matrix(data)
sol = trajeR(Y = data[, 2:6], A = data[, 7:11], degre = c(1,2), Model = "LOGIT", Method = "L")
sol
}
