% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.POIS}
\alias{print.Trajectory.POIS}
\title{Print POIS}
\usage{
\method{print}{Trajectory.POIS}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. . An object of class "\code{Trajectory.POIS}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print method for an object of class "\code{Trajectory.POIS}".
}
\examples{
data <- read.csv(system.file("extdata", "POIS2gr.csv", package = "trajeR"))
data <- as.matrix(data)
sol <- trajeR(
  Y = data[, 2:6], A = data[, 7:11],
  degre = c(2, 2), Model = "POIS", Method = "L", hessian = FALSE
)
sol
}
