% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrintTrajeR.R
\name{print.Trajectory.ZIP}
\alias{print.Trajectory.ZIP}
\title{Print ZIP}
\usage{
\method{print}{Trajectory.ZIP}(x, ...)
}
\arguments{
\item{x}{Trajectory's object. An object of class "\code{Trajectory.ZIP}".}

\item{...}{optional parameters}
}
\value{
The print of Obj.
}
\description{
Print method for an object of class "\code{Trajectory.ZIP}".
}
\examples{
data <- read.csv(system.file("extdata", "ZIP2gr.csv", package = "trajeR"))
data <- as.matrix(data)
sol <- trajeR(
  Y = data[, 2:6], A = data[, 7:11],
  degre = c(1, 2), degre.nu = c(1, 1), Model = "ZIP", Method = "L"
)
sol
}
