% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_adequacy.R
\name{propAssign}
\alias{propAssign}
\title{Assignment proportion}
\usage{
propAssign(sol, Y, A)
}
\arguments{
\item{sol}{Trajectory's object. An object of type trajectory.}

\item{Y}{Matrix. A matrix containing the variables in the model.}

\item{A}{Matrix. A matrix containing the time variable data.}
}
\value{
A vector of real. The  proportion.
}
\description{
Calculate the proportion of individuals in a given group. That is the ratio of
the number of individuals in one group and all the individuals.
}
\examples{
\dontrun{
load("data/dataNORM01.RData")
solL = trajeR(data[,1:5], data[,6:10], ng = 3, degre=c(2,2,2), 
              Model="CNORM", Method = "L", ssigma = FALSE, 
              hessian = TRUE)
propAssign(sol, Y = data[, 1;5], A = data[, 6:10])
}
}
